# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para el sistema de monitoreo de instancias Proxmox El Centinela.\n**Flujo de autenticación**: Login → `/auth/login` → obtener `jwtTemporal` → `/auth/2fa/verify` → obtener `accessToken` → usar en el botón **Authorize** de esta UI.",
        "title": "El Centinela API",
        "contact": {
            "name": "Soporte El Centinela"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida la contraseña actual y aplica la nueva. Reglas de complejidad: entre 8 y 12 caracteres, al menos una mayúscula, un número y un carácter especial (!@#$%^\u0026*-_=+). Si la contraseña era temporal (`cambioContrasenaRequerido=true`), este cambio limpia ese flag y desbloquea el acceso al resto de la plataforma.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mi cuenta"
                ],
                "summary": "Cambiar mi contraseña",
                "parameters": [
                    {
                        "description": "Contraseña actual y nueva",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CambiarContrasenaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Formato inválido, contraseña actual incorrecta, nueva igual a la actual o no cumple las reglas de complejidad",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "PASSWORD_CHANGE_REQUIRED — solo este endpoint y logout son accesibles mientras el flag esté activo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "description": "Invalida la sesión asociada al refresh token en la cookie `centinela_refresh`. Emite la eliminación de la cookie y responde 204.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión (logout)",
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    },
                    "401": {
                        "description": "Refresh token inválido, sesión ya cerrada o cookie ausente",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Mensaje descriptivo del error."
                }
            }
        },
...
        "ports.CambiarContrasenaInput": {
            "type": "object",
            "required": [
                "contrasenaActual",
                "contrasenaNueva"
            ],
            "properties": {
                "contrasenaActual": {
                    "type": "string"
                },
                "contrasenaNueva": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 8
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de acceso definitivo. Formato: **Bearer \u0026lt;accessToken\u0026gt;**. Obtené el token completando el flujo: `POST /auth/login` → `POST /auth/2fa/verify`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerPreAuth": {
            "description": "JWT temporal pre-2FA. Formato: **Bearer \u0026lt;jwtTemporal\u0026gt;**. Obtené el token de `POST /auth/login`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}